requirements.txt
langgraph>=0.2.60
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0
rich>=13.0
tenacity>=8.2.0
